Moved MPDUriMapper to own file and updated changelog
