+++ mopidy/mpd/actor.py
@@ -6,7 +6,7 @@ import pykka
 
 from mopidy import exceptions, zeroconf
 from mopidy.core import CoreListener
-from mopidy.mpd import dispatcher, session
+from mopidy.mpd import session, uri_mapper
 from mopidy.utils import encoding, network, process
 
 logger = logging.getLogger(__name__)
@@ -18,7 +18,7 @@ class MpdFrontend(pykka.ThreadingActor, CoreListener):
 
         self.hostname = network.format_hostname(config['mpd']['hostname'])
         self.port = config['mpd']['port']
-        self.uri_map = dispatcher.MpdUriMapper(core)
+        self.uri_map = uri_mapper.MpdUriMapper(core)
 
         self.zeroconf_name = config['mpd']['zeroconf']
         self.zeroconf_service = None
