+++ tests/mpd/protocol/__init__.py
@@ -8,7 +8,7 @@ import pykka
 
 from mopidy import core
 from mopidy.backend import dummy
-from mopidy.mpd import dispatcher, session
+from mopidy.mpd import session, uri_mapper
 
 
 class MockConnection(mock.Mock):
@@ -35,7 +35,7 @@ class BaseTestCase(unittest.TestCase):
         self.backend = dummy.create_dummy_backend_proxy()
         self.core = core.Core.start(backends=[self.backend]).proxy()
 
-        self.uri_map = dispatcher.MpdUriMapper(self.core)
+        self.uri_map = uri_mapper.MpdUriMapper(self.core)
         self.connection = MockConnection()
         self.session = session.MpdSession(
             self.connection, config=self.get_config(), core=self.core,
